splitted handlers to files, added error handling
